% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_documentation.R
\name{print_documentation}
\alias{print_documentation}
\title{Generate Documentation for a Package}
\usage{
print_documentation(package_name)
}
\arguments{
\item{package_name}{Name of the package for which to generate function documentation.}
}
\value{
Generates a markdown file with the function documentation and converts it to PDF.
}
\description{
Generate Documentation for a Package
}
